att projeto/orcamento

+++ src/main/java/lad/sys/api/repository/ProjetoRepository.java
@@ -5,9 +5,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProjetoRepository extends JpaRepository<Projeto, Long> {
+public interface ProjetoRepository extends JpaRepository<Projeto, Integer> {
 
 	Page<Projeto> findAllByAtivoTrue(Pageable pageable);
 
-	Projeto getReferenceByIdAndAtivoTrue(Long id);
+	Projeto getReferenceByIdAndAtivoTrue(Integer id);
 }
